# src/routes/api_routes.ts
import express from 'express';
import { registerAgent, listAgents, getAgentById, updateAgent, removeAgent } from '../services/agentService';

const router = express.Router();

// Middleware to extract tenantId from headers or query
const tenantMiddleware = (req: any, res: any, next: any) => {
  req.tenantId = req.headers['x-tenant-id'] || req.query.tenantId || 'default';
  next();
};

// Placeholder middleware and handlers
const authMiddleware = (req: any, res: any, next: any) => next();

const createTask = (req: any, res: any) => res.status(501).send('Not Implemented');
const getTaskStatus = (req: any, res: any) => res.status(501).send('Not Implemented');
const listTasks = (req: any, res: any) => res.status(501).send('Not Implemented');
const cancelTask = (req: any, res: any) => res.status(501).send('Not Implemented');
const streamTaskLogs = (req: any, res: any) => res.status(501).send('Not Implemented');

const registerAgentHandler = async (req: any, res: any) => {
  try {
    const agent = await registerAgent(req.body);
    res.json(agent);
  } catch (error: any) {
    res.status(500).json({ error: error.message });
  }
};

const listAgentsHandler = async (req: any, res: any) => {
  try {
    const agents = await listAgents();
    res.json(agents);
  } catch (error: any) {
    res.status(500).json({ error: error.message });
  }
};

const getAgentHandler = async (req: any, res: any) => {
  try {
    const agent = await getAgentById(req.params.id);
    if (!agent) {
      res.status(404).json({ error: 'Agent not found' });
    } else {
      res.json(agent);
    }
  } catch (error: any) {
    res.status(500).json({ error: error.message });
  }
};

const updateAgentHandler = async (req: any, res: any) => {
  try {
    const agent = await updateAgent(req.params.id, req.body);
    if (!agent) {
      res.status(404).json({ error: 'Agent not found' });
    } else {
      res.json(agent);
    }
  } catch (error: any) {
    res.status(500).json({ error: error.message });
  }
};

const removeAgentHandler = async (req: any, res: any) => {
  try {
    await removeAgent(req.params.id);
    res.json({ success: true });
  } catch (error: any) {
    res.status(500).json({ error: error.message });
  }
};

const runGraphQuery = (req: any, res: any) => res.status(501).send('Not Implemented');
const getGraphSchema = (req: any, res: any) => res.status(501).send('Not Implemented');
const createGraphNode = (req: any, res: any) => res.status(501).send('Not Implemented');
const updateGraphNode = (req: any, res: any) => res.status(501).send('Not Implemented');
const deleteGraphNode = (req: any, res: any) => res.status(501).send('Not Implemented');

const healthCheck = (req: any, res: any) => res.json({ status: 'healthy' });
const getVersion = (req: any, res: any) => res.json({ version: '1.0.0' });
const serveDocs = (req: any, res: any) => res.status(501).send('Not Implemented');
const getConfig = (req: any, res: any) => res.status(501).send('Not Implemented');

const login = (req: any, res: any) => res.status(501).send('Not Implemented');
const refreshToken = (req: any, res: any) => res.status(501).send('Not Implemented');
const logout = (req: any, res: any) => res.status(501).send('Not Implemented');
const getProfile = (req: any, res: any) => res.status(501).send('Not Implemented');

// Orchestrator Task Management
router.post('/tasks', tenantMiddleware, authMiddleware, createTask);
router.get('/tasks/:id', tenantMiddleware, authMiddleware, getTaskStatus);
router.get('/tasks', tenantMiddleware, authMiddleware, listTasks);
router.post('/tasks/:id/cancel', tenantMiddleware, authMiddleware, cancelTask);
router.get('/tasks/:id/logs', tenantMiddleware, authMiddleware, streamTaskLogs);

// Agent Registry & Discovery
router.post('/agents', tenantMiddleware, authMiddleware, registerAgentHandler);
router.get('/agents', tenantMiddleware, authMiddleware, listAgentsHandler);
router.get('/agents/:id', tenantMiddleware, authMiddleware, getAgentHandler);
router.patch('/agents/:id', tenantMiddleware, authMiddleware, updateAgentHandler);
router.delete('/agents/:id', tenantMiddleware, authMiddleware, removeAgentHandler);

// Knowledge Graph Queries
router.post('/graph/query', tenantMiddleware, authMiddleware, runGraphQuery);
router.get('/graph/schema', tenantMiddleware, authMiddleware, getGraphSchema);
router.post('/graph/node', tenantMiddleware, authMiddleware, createGraphNode);
router.patch('/graph/node/:id', tenantMiddleware, authMiddleware, updateGraphNode);
router.delete('/graph/node/:id', tenantMiddleware, authMiddleware, deleteGraphNode);

// System/Utility Endpoints
router.get('/health', healthCheck);
router.get('/version', getVersion);
router.get('/docs', serveDocs);
router.get('/config', tenantMiddleware, authMiddleware, getConfig);

// Auth Routes
// Auth Routes
router.post('/auth/login', login);
router.post('/auth/refresh', refreshToken);
router.post('/auth/logout', logout);
router.get('/auth/profile', tenantMiddleware, authMiddleware, getProfile);

export default router;